pyproject.toml
[tool.poetry]
name = "django-drf"
version = "0.1.0"
description = ""
authors = ["alexandr <dmitrymaslov2016@mai.ru>"]
readme = "README.md"
packages = [{include = "django_drf"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.3"
djangorestframework = "^3.14.0"
pillow = "^10.0.0"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.6"
django-filter = "^23.2"
pre-commit = "^3.3.3"
pytest = "^7.4.0"
djangorestframework-simplejwt = "^5.2.2"
drf-yasg = "^1.21.7"
stripe = "^5.5.0"
requests = "^2.31.0"
celery = "^5.3.1"
redis = "^4.6.0"
django-celery-beat = "^2.5.0"
flower = "^2.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
